RSpec tests now run properly
Rspec was failing due to broken queries to the database because the valid_attributes hash didn't contain all of the necessary default values needed to query the database. For example, creating a Section was failing in the tests because no value for the number attribute was specified, so it was null, and the database won't allow a null value when creating a Section.