Continued working on Assignment pages
TODO: Need to get rid of the courses/:id/sections/:id/assignments/:id nested format and instead only show one of them at once. Sections should never even be shown in the URL since a user can't change what section they are in.

To view all of their courses, the URL courses/ should do just that and take the user to a page modified by the professor at courses/:id. Listed assignments for the course should be at courses/:id/assignments.

If a user wants to view all of their assignments, they should go to assignments/, and if a user wants to view the assignments for a particular course, the URL should be courses/:id/assignments (or assignments?course=:id).

If the user is viewing an assignment, the URL should be courses/:id/assignments/:id. However, it would be nice if the URL could have a singular model name like assignment/:id and have the app query for the course ID etc.