Database updated
The database was upgraded to support the Assignments Upload page and the Attendance page which have now both been integrated into the master branch. Also, the teacher gradebook and the normal gradebook are currently broken since the Ruby code needs to be upgraded to work with the correct database tables.