Attendance Progress
Working on the attendance page. Everything will be done on the index page of the attendance resource. If the parameters section and date are available, then the attendance for the given section and day will be shown for the professor to work with. However, if there are no params and the URL is just /attendances/, then all the sections the professor is currently a professor of will be shown to the professor to click on.