Attendance Updates
Added attendances in seeds file. The attendance home page now shows the available courses, their sections, and a link to set the attendance for the section. I still need to only display the sections for the logged in user which will require some additions to seeds.rb

The actual attendance page for a section now shows the students in the given course and their attendance information. However, I still need to make this modifiable by the professor.