Added Roles, Attendance Updates
Added the following roles: Admin, Professor, TA, Student, and Observer.

Made some good progres with the Attendance queries with the new sections_users_roles many to many to many table. I can now query for the students, tas, and professors of a class and display them. The next thing is to allow attendance to actually be taken.