Attendance Updates
Updated seeds.rb so every section is populated with a professor, a TA, and students. Also, so not every user is in every section, users with even IDs are inserted into sections with even IDs and vice versa. So, you can login as a professor and you will only see the sections you are a professor of on the attendance page, and you won't see anything if you are a student.

Below are the test accounts we should be using from now on:

admin@whiteboard.com
prof1@whiteboard.com
prof2@whiteboard.com
ta1@whiteboard.com
ta2@whiteboard.com
student1@whiteboard.com
student2@whiteboard.com
student3@whiteboard.com

The password is 'password' for all of them.