Assignment Updates
Professors can now add new assignments for a section as well as update them. Also fixed a bug where each section of a course was showing for a professor, even if that professor didn't teach the given course.

Had to update the URL for the upload assignment page from assignments/1/edit (now for professor assignment updating) to assignments/1/files.