Attendance Updates & Texas University
Fixed a bug with the attendance where the dates were out of sync since I wasn't pulling from the date in the URL in the attendance_controller and so at certain times of the day the date would be different thus causing the professor to get an error message when attempting to update a student's attendance since their attendance was never created in the database and didn't exist.

Also updated the attendance failure error message so it's not an ugly alert message, but a pretty temporary bold and red message that fades out over three seconds time.

For Universities, I added UT professors and students as well as some courses and assignments.