Appointment & Assignment Dates Fixed
Whenever you set an apointment with a message or create an assignment and set the due date, the date and time should now be correct. This will need to be tested on the hosting device however, since results seems to differ between timestamp with and without time zones. I've ended up using a time zone timestamp, but I don't think it is any different from a timestamp without a time zone since I added the -00:00 parameter to the Time.new method.